+++ src/App.vue
@@ -1,8 +1,9 @@
 <template>
-  <tyh-rate v-model="value1" />
+  <tyh-switch />
+  <tyh-switch v-model="value1" />
 </template>
 
 <script setup>
 import { ref } from 'vue'
-const value1 = ref(22)
+const value1 = ref(true)
 </script>