+++ src/packages/tyhUi/packages/select/TyhSelect.vue
@@ -1,10 +1,16 @@
 <template>
-  <select class="tyh-select" v-model="modelValue" :name="name" @input="input">
+  <select
+    v-if="$slots.default"
+    class="tyh-select"
+    v-model="modelValue"
+    :name="name"
+    @input="input"
+  >
     <slot />
   </select>
 </template>
 
-<script setup>
+<script lang="ts" setup>
 defineProps({
   modelValue: {
     type: String,
@@ -13,5 +19,5 @@ defineProps({
   name: String
 })
 const emit = defineEmits(['update:modelValue'])
-const input = e => emit('update:modelValue', e.target.value)
+const input = (e: any): void => emit('update:modelValue', e.target.value)
 </script>