+++ src/packages/tyhUi/packages/skeleton/TyhSkeleton.vue
@@ -11,7 +11,7 @@
   />
 </template>
 
-<script setup>
+<script lang="ts" setup>
 defineProps({
   round: Boolean,
   width: String,